refactor: Split out typos-dict
