+++ benches/file.rs
@@ -12,7 +12,7 @@ fn process_empty(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::EMPTY).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
@@ -26,7 +26,7 @@ fn process_no_tokens(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::NO_TOKENS).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
@@ -40,7 +40,7 @@ fn process_single_token(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::SINGLE_TOKEN).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
@@ -54,7 +54,7 @@ fn process_sherlock(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::SHERLOCK).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
@@ -68,7 +68,7 @@ fn process_code(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::CODE).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
@@ -82,7 +82,7 @@ fn process_corpus(b: &mut test::Bencher) {
     let sample_path = temp.child("sample");
     sample_path.write_str(data::CORPUS).unwrap();
 
-    let corrections = typos::BuiltIn::new();
+    let corrections = typos_dict::BuiltIn::new();
     let parser = typos::tokens::Parser::new();
     let checks = typos::checks::CheckSettings::new().build(&corrections, &parser);
     b.iter(|| checks.check_file(sample_path.path(), true, typos::report::print_silent));
