+++ typos-dict/src/lib.rs
@@ -1,0 +1,4 @@
+mod dict;
+mod dict_codegen;
+
+pub use crate::dict::*;
