+++ typos/Cargo.toml
@@ -15,13 +15,8 @@ edition = "2018"
 travis-ci = { repository = "epage/typos" }
 appveyor  = { repository = "epage/typos" }
 
-[features]
-# Support quickly iterating
-iterate_unstable = []
-
 [dependencies]
 failure = "0.1"
-phf = { version = "0.7", features = ["unicase"] }
 regex = "1.0"
 lazy_static = "1.2.0"
 serde = { version = "1.0", features = ["derive"] }
@@ -32,8 +27,3 @@ bstr = "0.2"
 log = "0.4"
 unicode-segmentation = "1.3.0"
 derive_more = "0.15.0"
-
-[build-dependencies]
-phf_codegen = "0.7"
-csv = "1.0"
-unicase = "1.1"
