+++ typos/src/lib.rs
@@ -1,5 +1,4 @@
 mod dict;
-mod dict_codegen;
 
 pub mod checks;
 pub mod report;
